+++ src/components/film-card/film-card.tsx
@@ -1,4 +1,4 @@
-import {FC} from 'react';
+import { FC, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { RoutePaths, getRoutePath } from '../../config/route';
 import { FilmCard as FilmCardType } from '../../types/film';
@@ -6,28 +6,31 @@ import { VideoPreview } from '../video-preview/video-preview';
 
 type FilmCardProps = {
   filmCard: FilmCardType;
-  active: boolean;
-  setActiveCard: React.Dispatch<React.SetStateAction<string | null>>;
 }
 
-const FilmCard: FC<FilmCardProps> = ({filmCard, setActiveCard, active}) => (
-  <article
-    className="small-film-card catalog__films-card"
-    onMouseOver={() => setActiveCard(filmCard.id)}
-    onMouseOut={() => setActiveCard(null)}
-  >
-    <VideoPreview
-      className="small-film-card__image"
-      isActive={active}
-      previewImage={filmCard.previewImage}
-      videoLink={filmCard.previewVideoLink}
-      alt={filmCard.name}
-      videoTimeout={1000}
-    />
-    <h3 className="small-film-card__title">
-      <Link className="small-film-card__link" to={getRoutePath(RoutePaths.Film, {id: filmCard.id})}>{filmCard.name}</Link>
-    </h3>
-  </article>
-);
+const FilmCard: FC<FilmCardProps> = ({filmCard}) => {
+  const [active, setActive] = useState<boolean | null>(false);
+
+  return (
+    <article
+      className="small-film-card catalog__films-card"
+      onMouseOver={() => setActive(true)}
+      onMouseOut={() => setActive(null)}
+    >
+      <VideoPreview
+        className="small-film-card__image"
+        isActive={active ?? false}
+        previewImage={filmCard.previewImage}
+        videoLink={filmCard.previewVideoLink}
+        alt={filmCard.name}
+        videoTimeout={1000}
+      />
+      <h3 className="small-film-card__title">
+        <Link className="small-film-card__link" to={getRoutePath(RoutePaths.Film, {id: filmCard.id})}>{filmCard.name}</Link>
+      </h3>
+    </article>
+  );
+};
 
 export default FilmCard;
+
