+++ src/components/film-list/film-list.tsx
@@ -1,4 +1,4 @@
-import {FC, useState} from 'react';
+import { FC } from 'react';
 import FilmCard from '../film-card/film-card';
 import { FilmCard as FilmCardType } from '../../types/film/index';
 
@@ -6,21 +6,15 @@ interface FilmListProps {
   filmCards: FilmCardType[];
 }
 
-const FilmList: FC<FilmListProps> = ({filmCards}) => {
-  const [activeCard, setActiveCard] = useState<string | null>(null);
-
-  return (
-    <div className="catalog__films-list">
-      {filmCards.map((filmCard) => (
-        <FilmCard
-          key={filmCard.id}
-          filmCard={filmCard}
-          active={activeCard === filmCard.id}
-          setActiveCard={setActiveCard}
-        />
-      ))}
-    </div>
-  );
-};
+const FilmList: FC<FilmListProps> = ({filmCards}) => (
+  <div className="catalog__films-list">
+    {filmCards.map((filmCard) => (
+      <FilmCard
+        key={filmCard.id}
+        filmCard={filmCard}
+      />
+    ))}
+  </div>
+);
 
 export default FilmList;
